third_party/WebKit/LayoutTests/fast/workers/worker-onerror-02-expected.txt

This tests that unhandled exceptions trigger 'worker.onerror'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Page-level worker.onerror handler triggered:
PASS errorEvent.message is "Uncaught ReferenceError: foo is not defined"
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 2
PASS errorEvent.colno is 9
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
